Introduce `BuildTreeActionExecutor` service
