+++ platforms/core-runtime/launcher/src/main/java/org/gradle/launcher/exec/BuildTreeLifecycleBuildActionExecutor.java
@@ -23,55 +23,103 @@ import org.gradle.internal.buildtree.BuildActionModelRequirements;
 import org.gradle.internal.buildtree.BuildActionRunner;
 import org.gradle.internal.buildtree.BuildModelParameters;
 import org.gradle.internal.buildtree.BuildModelParametersFactory;
+import org.gradle.internal.buildtree.BuildTreeActionExecutor;
 import org.gradle.internal.buildtree.BuildTreeState;
 import org.gradle.internal.buildtree.RunTasksRequirements;
 import org.gradle.internal.hash.HashCode;
 import org.gradle.internal.hash.Hashing;
 import org.gradle.internal.id.UniqueId;
 import org.gradle.internal.invocation.BuildAction;
+import org.gradle.internal.operations.BuildOperationContext;
+import org.gradle.internal.operations.BuildOperationDescriptor;
+import org.gradle.internal.operations.BuildOperationRunner;
+import org.gradle.internal.operations.CallableBuildOperation;
+import org.gradle.internal.operations.logging.LoggingBuildOperationProgressBroadcaster;
+import org.gradle.internal.operations.notify.BuildOperationNotificationValve;
+import org.gradle.internal.problems.failure.Failure;
 import org.gradle.internal.scopeids.id.BuildInvocationScopeId;
 import org.gradle.internal.service.ServiceRegistry;
-import org.gradle.internal.session.BuildSessionActionExecutor;
-import org.gradle.internal.session.BuildSessionContext;
 import org.gradle.internal.snapshot.ValueSnapshotter;
+import org.gradle.internal.work.WorkerLeaseService;
 import org.gradle.tooling.internal.provider.action.BuildModelAction;
 import org.gradle.tooling.internal.provider.action.ClientProvidedBuildAction;
 import org.gradle.tooling.internal.provider.action.ClientProvidedPhasedAction;
 import org.gradle.tooling.internal.provider.serialization.SerializedPayload;
+import org.jspecify.annotations.Nullable;
 
 import java.util.function.Supplier;
 
-/**
- * A {@link BuildActionExecutor} responsible for establishing the build tree for a single invocation of a {@link BuildAction}.
- */
-public class BuildTreeLifecycleBuildActionExecutor implements BuildSessionActionExecutor {
+public class BuildTreeLifecycleBuildActionExecutor implements BuildTreeActionExecutor {
+
+    private static final RunBuildBuildOperationType.Details DETAILS = new RunBuildBuildOperationType.Details() {
+    };
 
     private final BuildModelParametersFactory buildModelParametersFactory;
     private final BuildLayoutValidator buildLayoutValidator;
     private final ValueSnapshotter valueSnapshotter;
     private final InternalOptions options;
+    private final WorkerLeaseService workerLeaseService;
+    private final BuildOperationRunner buildOperationRunner;
+    private final LoggingBuildOperationProgressBroadcaster loggingBuildOperationProgressBroadcaster;
+    private final BuildOperationNotificationValve buildOperationNotificationValve;
 
     public BuildTreeLifecycleBuildActionExecutor(
         BuildModelParametersFactory modelParametersFactory,
         BuildLayoutValidator buildLayoutValidator,
         ValueSnapshotter valueSnapshotter,
-        InternalOptions options
+        InternalOptions options,
+        WorkerLeaseService workerLeaseService,
+        BuildOperationRunner buildOperationRunner,
+        LoggingBuildOperationProgressBroadcaster loggingBuildOperationProgressBroadcaster,
+        BuildOperationNotificationValve buildOperationNotificationValve
     ) {
         this.buildModelParametersFactory = modelParametersFactory;
         this.buildLayoutValidator = buildLayoutValidator;
         this.valueSnapshotter = valueSnapshotter;
         this.options = options;
+        this.workerLeaseService = workerLeaseService;
+        this.buildOperationRunner = buildOperationRunner;
+        this.loggingBuildOperationProgressBroadcaster = loggingBuildOperationProgressBroadcaster;
+        this.buildOperationNotificationValve = buildOperationNotificationValve;
     }
 
     @Override
-    public BuildActionRunner.Result execute(BuildAction action, BuildSessionContext buildSession) {
+    public BuildActionRunner.Result runBuildTreeAction(BuildAction action, ServiceRegistry buildSessionServices) {
+        return workerLeaseService.runAsWorkerThread(() -> runAsBuildOperation(action, buildSessionServices));
+    }
+
+    private BuildActionRunner.Result runAsBuildOperation(BuildAction action, ServiceRegistry buildSessionServices) {
+        buildOperationNotificationValve.start();
+        try {
+            return buildOperationRunner.call(new CallableBuildOperation<BuildActionRunner.Result>() {
+                @Override
+                public BuildActionRunner.Result call(BuildOperationContext buildOperationContext) {
+                    loggingBuildOperationProgressBroadcaster.rootBuildOperationStarted();
+                    BuildActionRunner.Result result = runBuildTreeLifecycle(action, buildSessionServices);
+                    buildOperationContext.setResult(new DefaultRunBuildResult(result));
+                    if (result.getBuildFailure() != null) {
+                        buildOperationContext.failed(result.getBuildFailure());
+                    }
+                    return result;
+                }
+
+                @Override
+                public BuildOperationDescriptor.Builder description() {
+                    return BuildOperationDescriptor.displayName("Run build").details(DETAILS);
+                }
+            });
+        } finally {
+            buildOperationNotificationValve.stop();
+        }
+    }
+
+    private BuildActionRunner.Result runBuildTreeLifecycle(BuildAction action, ServiceRegistry buildSessionServices) {
         BuildActionRunner.Result result = null;
         try {
             buildLayoutValidator.validate(action.getStartParameter());
 
             BuildActionModelRequirements actionRequirements = buildActionModelRequirementsFor(action);
             BuildModelParameters buildModelParameters = buildModelParametersFactory.parametersForRootBuildTree(actionRequirements, options);
-            ServiceRegistry buildSessionServices = buildSession.getServices();
             BuildInvocationScopeId buildInvocationScopeId = new BuildInvocationScopeId(UniqueId.generate());
             try (BuildTreeState buildTree = new BuildTreeState(buildSessionServices, actionRequirements, buildModelParameters, buildInvocationScopeId)) {
                 // assign instead of return to allow combining build failures with cleanup failures below
@@ -106,5 +154,19 @@ public class BuildTreeLifecycleBuildActionExecutor implements BuildSessionAction
     private Supplier<HashCode> payloadHashProvider(Object payload) {
         ValueSnapshotter valueSnapshotter = this.valueSnapshotter;
         return () -> Hashing.hashHashable(valueSnapshotter.snapshot(payload));
+    }
+
+    private static class DefaultRunBuildResult implements RunBuildBuildOperationType.Result {
+        private final BuildActionRunner.Result result;
+
+        public DefaultRunBuildResult(BuildActionRunner.Result result) {
+            this.result = result;
+        }
+
+        @Override
+        @Nullable
+        public Failure getFailure() {
+            return result.getRichBuildFailure();
+        }
     }
 }
