--- platforms/core-runtime/launcher/src/main/java/org/gradle/launcher/exec/RunAsWorkerThreadBuildActionExecutor.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2021 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.launcher.exec;
-
-import org.gradle.internal.buildtree.BuildActionRunner;
-import org.gradle.internal.invocation.BuildAction;
-import org.gradle.internal.session.BuildSessionActionExecutor;
-import org.gradle.internal.session.BuildSessionContext;
-import org.gradle.internal.work.WorkerLeaseService;
-
-public class RunAsWorkerThreadBuildActionExecutor implements BuildSessionActionExecutor {
-    private final BuildSessionActionExecutor delegate;
-    private final WorkerLeaseService workerLeaseService;
-
-    public RunAsWorkerThreadBuildActionExecutor(WorkerLeaseService workerLeaseService, BuildSessionActionExecutor delegate) {
-        this.delegate = delegate;
-        this.workerLeaseService = workerLeaseService;
-    }
-
-    @Override
-    public BuildActionRunner.Result execute(BuildAction action, BuildSessionContext context) {
-        return workerLeaseService.runAsWorkerThread(() -> delegate.execute(action, context));
-    }
-}
